+++ app/src/main/res/layout/activity_main.xml
@@ -7,8 +7,8 @@
     android:layout_height="match_parent">
 
     <fragment
-        android:id="@+id/calcurationFragment"
-        android:name="com.example.bmiapplication.CalcurationFragment"
+        android:id="@+id/calculationFragment"
+        android:name="com.example.bmiapplication.CalculationFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
